jpeg: handle embeded subimage/thumbnail
The original way of calculating the size of a jpeg image is like this:
first find a SOI segments which indicates the start of an image,
then try to find the second one, if the second one is found,
then subtract their offset to get the size, if the second one is not found,
then the size is the same as the file size.

So if Appn segment contains a subimage, the second SOI is the subimage's SOI,
not the next jpeg image's.

Here the program tries to find the corresponding EOI of the SOI to
get the size of a jpeg image. It is tested successfully both for jpeg and M-JPEG.

Signed-off-by: Junfeng Xu <jun.feng.xu@intel.com>