MVC dec: Handle the prefix NAL unit correctly.
Cache the prefix_nal until we finish the parsing of next nal unit.
If the nal unit following the prefix_nal unit is a SLICE/SLICE_IDR nal then,
the prefix_nal should be considered as the part of it's associated
nal unit. Likewise we should skip the prefix_nal and associated_nal if the
associated nal is not a SLICE/SLICE_IDR.

committer: Sreerenj Balachandran <sreerenj.balachandran@intel.com>
Signed-off-by: Li Xiaowei <xiaowei.a.li@intel.com>